+++ astro.config.mjs
@@ -7,6 +7,7 @@ import { unified } from "@astrojs/markdown-remark";
 import tailwindcss from "@tailwindcss/vite";
 import remarkToc from "remark-toc";
 import remarkCollapse from "remark-collapse";
+import { remarkEmbeds } from "./src/utils/remarkEmbeds.mjs";
 import { remarkLazyLoadImages } from "./src/utils/remarkLazyLoadImages.mjs";
 import { SITE } from "./src/config";
 import AstroPWA from "@vite-pwa/astro";
@@ -21,6 +22,7 @@ export default defineConfig({
         remarkToc,
         [remarkCollapse, { test: "Table of contents" }],
         remarkLazyLoadImages,
+        remarkEmbeds,
       ],
     }),
     shikiConfig: {